+++ kcauto-kai.sikuli/combat.sikuli/combat.py
@@ -44,7 +44,7 @@ class CombatModule(object):
         self.map = MapData(
             self.config.combat['map'], self.regions, self.config)
         self.current_position = [0, 0]
-        self.current_node = None
+        self.current_node = self.map.unknown_node
         self.nodes_run = []
 
         self.lbas = (
@@ -106,7 +106,7 @@ class CombatModule(object):
         self.stats.increment_combat_attempted()
 
         if not self._select_combat_map():
-            # LBAS fatigue check failed; cancel sortie
+            # Port check or LBAS fatigue check failed; cancel sortie
             return False
 
         if self._conduct_pre_sortie_checks():
@@ -154,11 +154,12 @@ class CombatModule(object):
 
     def _select_combat_map(self):
         """Method that goes through the menu and chooses the specified map to
-        sortie to. LBAS checks are also resolved at this point.
+        sortie to. Non-event port checks and LBAS checks are also resolved at
+        this point.
 
         Returns:
             bool: True if the combat map is successfully chosen and started,
-                False if an LBAS check failed
+                False if a port or LBAS check failed
         """
         Util.rejigger_mouse(self.regions, 'top')
 
@@ -209,7 +210,10 @@ class CombatModule(object):
                 Util.kc_sleep(2)
             Util.wait_and_click(self.kc_region, 'c_world_{}-{}.png'.format(
                 self.map.world, self.map.subworld))
-        Util.wait_and_click(self.regions['lower_right'], 'sortie_select.png')
+        self.regions['lower_right'].wait('sortie_select.png')
+        if self._delay_from_port_check('select'):
+            return False
+        Util.check_and_click(self.regions['lower_right'], 'sortie_select.png')
         Util.rejigger_mouse(self.regions, 'top')
         return True
 
@@ -272,15 +276,16 @@ class CombatModule(object):
             self.set_next_combat_time()
             cancel_sortie = True
 
-        if ('PortCheck' in self.config.combat['misc_options'] or
-                self.map.world == 'event'):
-            port_full_notice = (
-                'warning_port_full_event.png'
-                if self.map.world == 'event' else 'warning_port_full.png')
-            if self.regions['lower'].exists(port_full_notice):
-                Util.log_warning("Canceling combat sortie: port is full.")
-                self.set_next_combat_time({'minutes': 15})
-                cancel_sortie = True
+        if self.dmg['repair'] > 0:
+            Util.log_warning(
+                "Canceling combat sortie: {:d} ships are being repaired."
+                .format(self.dmg['repair']))
+            self.set_next_combat_time()
+            self.primary_fleet.force_check_repair = True
+            cancel_sortie = True
+
+        if self._delay_from_port_check('sortie'):
+            cancel_sortie = True
 
         if cancel_sortie:
             return False
@@ -316,6 +321,37 @@ class CombatModule(object):
             return (needs_resupply, fleet_damages, fleet_fatigue)
         return (needs_resupply, fleet_damages, {})
 
+    def _delay_from_port_check(self, context):
+        """Checks to see if the sortie needs to be delayed due to a full port
+        and the config that checks for this, or the map being an event map.
+
+        Args:
+            context (str): 'select' or 'sortie', depending on which screen the
+                warning messages are being checked at
+
+        Returns:
+            bool: True if the sortie needs to be cancelled and delayed, False
+                otherwise
+        """
+        full_port = False
+        if ('PortCheck' in self.config.combat['misc_options'] or
+                self.map.world == 'event'):
+            # this logic may need to be reworked; the event-time messaging
+            # is a bit uncertain at the moment, since this is being revised not
+            # during an event
+            if context == 'select' and self.map.world != 'event':
+                if self.regions['right'].exists('warning_port_full.png'):
+                    full_port = True
+            elif context == 'sortie' and self.map.world == 'event':
+                if self.regions['lower'].exists('warning_port_full_event.png'):
+                    full_port = True
+        if full_port:
+            Util.log_warning("Canceling combat sortie: port is full.")
+            self.set_next_combat_time({'minutes': 15})
+            return True
+        else:
+            return False
+
     def _run_combat_logic(self):
         """Method that contains the logic and fires off necessary child methods
         for resolving anything combat-related. Includes LBAS node assignment,
@@ -334,15 +370,14 @@ class CombatModule(object):
 
         # primary combat loop
         sortieing = True
+        self.current_position = [0, 0]
+        self.current_node = self.map.unknown_node
+        self.current_node.reset_unknown_node()
         self.nodes_run = []
         disable_combat = False
         post_combat_screens = []
         while sortieing:
             at_node, dialogue_click = self._run_loop_between_nodes()
-
-            # stop the background observer if no longer on the map screen
-            if self.config.combat['engine'] == 'live':
-                self.observeRegion.stopObserver()
 
             if at_node:
                 # arrived at combat node
@@ -525,6 +560,7 @@ class CombatModule(object):
                         'formation_combinedfleet_1.png')):
                 # check for both single fleet and combined fleet formations
                 # since combined fleets can have single fleet battles
+                self._stop_fleet_observer()
                 self._print_current_node()
                 formations = self._resolve_formation()
                 for formation in formations:
@@ -552,6 +588,7 @@ class CombatModule(object):
                     or self.fast_kc_region.exists('combat_nb_fight.png')):
                 # post-combat or night battle select without selecting a
                 # formation
+                self._stop_fleet_observer()
                 self._print_current_node()
                 Util.rejigger_mouse(self.regions, 'lbas')
                 at_node = True
@@ -559,9 +596,11 @@ class CombatModule(object):
             elif self.regions['lower_right_corner'].exists(
                     'combat_flagship_dmg.png'):
                 # flagship retreat
+                self._stop_fleet_observer()
                 return (False, False)
             elif self.regions['lower_right_corner'].exists('next_alt.png'):
                 # resource node end
+                self._stop_fleet_observer()
                 return (False, False)
 
     def _run_loop_during_battle(self):
@@ -610,16 +649,11 @@ class CombatModule(object):
             fleet_match.y + fleet_match.h - self.kc_region.y
         ]
 
+        self.current_node = self.map.find_node_by_pos(*self.current_position)
         # debug console print for the observer's found position of the fleet
         """
-        print(
-            "{}, {} ({})".format(
-                self.current_position[0], self.current_position[1],
-                fleet_match))
+        print("{} {}".format(self.current_position, self.current_node))
         """
-        matched_node = self.map.find_node_by_pos(*self.current_position)
-        self.current_node = (
-            matched_node if matched_node is not None else self.current_node)
         event.repeat()
 
     def _update_fleet_position_once(self):
@@ -635,16 +669,11 @@ class CombatModule(object):
             fleet_match.y + fleet_match.h - self.kc_region.y
         ]
 
+        self.current_node = self.map.find_node_by_pos(*self.current_position)
         # debug console print for the method's found position of the fleet
         """
-        print(
-            "{}, {} ({})".format(
-                self.current_position[0], self.current_position[1],
-                fleet_match))
+        print("{} {}".format(self.current_position, self.current_node))
         """
-        matched_node = self.map.find_node_by_pos(*self.current_position)
-        self.current_node = (
-            matched_node if matched_node is not None else self.current_node)
         Util.log_msg("Fleet at node {}.".format(self.current_node))
 
     def _increment_nodes_run(self):
@@ -960,6 +989,7 @@ class CombatFleet(Fleet):
         self.damage_counts = {
             'repair': 0
         }
+        self.force_check_repair = False
         self.damaged_fcf_retreat_count = 0
         self.fatigue = {}
 
@@ -1146,8 +1176,10 @@ class CombatFleet(Fleet):
         """
         self.fatigue[mode] = (
             True
-            if (region.exists(Pattern('ship_state_fatigue_{}.png'.format(mode))
-                .similar(Globals.FATIGUE_SIMILARITY)))
+            if (region.exists(
+                Pattern('ship_state_fatigue_{}.png'.format(mode)).similar(
+                    Globals.FATIGUE_SIMILARITY
+                ), 1.5))
             else False)
 
     @staticmethod