+++ kcauto-kai.sikuli/repair.sikuli/repair.py
@@ -2,7 +2,7 @@ from sikuli import Pattern, Location
 from datetime import datetime, timedelta
 from kca_globals import Globals
 from combat import CombatFleet
-from nav import Nav
+from nav import Nav, NavList
 from util import Util
 
 
@@ -23,6 +23,11 @@ class RepairModule(object):
         self.kc_region = self.regions['game']
         self.fleets = fleets
         self.combat = combat
+
+        self.ship_count = 1
+        self.ship_page_count = 1
+        self.ship_last_page_count = 1
+        self.current_shiplist_page = 1
         self.repair_slots = 0
         self.repair_timers = []
 
@@ -30,6 +35,7 @@ class RepairModule(object):
         """Method to navigate to the repair menu.
         """
         Nav.goto(self.regions, 'repair')
+        self.current_shiplist_page = 1
 
     def check_need_to_repair(self):
         """Method to check whether or not ships need to be repaired in the
@@ -40,6 +46,13 @@ class RepairModule(object):
         """
         self._remove_old_timers()
         for fleet_id, fleet in self.fleets.items():
+            # first check if there is a forced check
+            if fleet.force_check_repair:
+                # immediately unset this flag so it does not cause false
+                # positives later on
+                fleet.force_check_repair = False
+                return True
+            # then check against damage counts
             if fleet.get_damage_counts_at_threshold(
                     self.config.combat['repair_limit']) > 0:
                 if (len(self.repair_timers) == self.repair_slots and
@@ -55,9 +68,10 @@ class RepairModule(object):
         menu.
         """
         Util.log_msg("Begin repairing fleets.")
+        self._set_shiplist_counts()
 
         # find busy docks and resolve existing repair timers
-        self.repair_timers = []
+        self.repair_timers = []  # clear previously tracked timers
         dock_busy_matches = Util.findAll_wrapper(
             self.kc_region, 'dock_timer.png')
         dock_busy_count = 0
@@ -96,6 +110,8 @@ class RepairModule(object):
                 else:
                     # no empty docks, so just exit out of loop
                     break
+        if len(self.repair_timers) > dock_busy_count:
+            self._update_combat_next_sortie_time()
 
     def _conduct_repair(self):
         """Method that chooses an empty dock, chooses a ship, toggles the
@@ -133,7 +149,7 @@ class RepairModule(object):
                 self.stats.increment_buckets_used()
                 self.kc_region.wait('dock_empty.png')
             else:
-                self._update_combat_next_sortie_time(repair_timer)
+                self._add_to_repair_timers(repair_timer)
                 self.regions['lower_right'].waitVanish('page_prev.png')
                 Util.kc_sleep(1)
         Util.kc_sleep()
@@ -161,34 +177,54 @@ class RepairModule(object):
         valid_damages = CombatFleet.get_damages_at_threshold(
             self.config.combat['repair_limit'])
 
-        for fleet_marker in fleet_markers:
-            fleet_id = int(fleet_marker[17])  # infer from filename
-            fleet_instance = self.fleets[fleet_id]
+        while self.current_shiplist_page <= self.ship_page_count:
+            for fleet_marker in fleet_markers:
+                fleet_id = int(fleet_marker[17])  # infer from filename
+                fleet_instance = self.fleets[fleet_id]
 
-            if fleet_instance.get_damage_counts_at_threshold(
-                    self.config.combat['repair_limit']) == 0:
-                # if this fleet has no longer has ships that need repair,
-                # don't search for its marker
-                continue
+                if fleet_instance.get_damage_counts_at_threshold(
+                        self.config.combat['repair_limit']) == 0:
+                    # if this fleet has no longer has ships that need repair,
+                    # don't search for its marker
+                    continue
 
-            ship_matches = Util.findAll_wrapper(
-                self.regions['repair_shiplist_fleet_markers'],
-                Pattern(fleet_marker).similar(0.9))
-            for ship_match in ship_matches:
-                target_region = ship_match.offset(Location(342, 0)).nearby(5)
+                ship_matches = Util.findAll_wrapper(
+                    self.regions['repair_shiplist_fleet_markers'],
+                    Pattern(fleet_marker).similar(0.9))
+                for ship_match in ship_matches:
+                    target_region = ship_match.offset(
+                        Location(342, 0)).nearby(5)
+                    for damage in valid_damages:
+                        if fleet_instance.damage_counts[damage] == 0:
+                            # if no ships in this fleet are at this damage
+                            # state, don't search for it
+                            continue
+
+                        damage_icon = 'repairlist_dmg_{}.png'.format(damage)
+                        if Util.check_and_click(
+                                target_region, damage_icon,
+                                Globals.EXPAND['repair_list']):
+                            fleet_instance.damage_counts[damage] -= 1
+                            fleet_instance.damage_counts['repair'] += 1
+                            return True
+            if self.current_shiplist_page < self.ship_page_count:
+                # check if there were even any damaged ships on the page
+                damage_exists = False
                 for damage in valid_damages:
-                    if fleet_instance.damage_counts[damage] == 0:
-                        # if no ships in this fleet are at this damage state,
-                        # don't search for it
-                        continue
-
                     damage_icon = 'repairlist_dmg_{}.png'.format(damage)
-                    if Util.check_and_click(
-                            target_region, damage_icon,
-                            Globals.EXPAND['repair_list']):
-                        fleet_instance.damage_counts[damage] -= 1
-                        fleet_instance.damage_counts['repair'] += 1
-                        return True
+                    if self.regions['right'].exists(damage_icon):
+                        damage_exists = True
+                        break
+                if damage_exists:
+                    # did not select a ship but damaged ships exist; go to next
+                    # page
+                    self._navigate_to_shiplist_page(
+                        self.current_shiplist_page + 1)
+                else:
+                    # no more ships of valid damage state exists in list;
+                    # return to the first page
+                    self._navigate_to_shiplist_page('first')
+                    return False
         return False
 
     def _pick_any_ship(self):
@@ -209,6 +245,35 @@ class RepairModule(object):
                 return True
         return False
 
+    def _set_shiplist_counts(self):
+        """Method that sets the ship-list related internal counts based on the
+        number of ships in the port.
+        """
+        self.ship_count, self.ship_page_count, self.ship_last_page_count = (
+            Util.get_shiplist_counts(self.regions))
+        Util.log_msg("Detecting {} ships across {} pages.".format(
+            self.ship_count, self.ship_page_count))
+
+    def _navigate_to_shiplist_page(self, target_page):
+        """Wrapper method that navigates the shiplist to the specified target
+        page from the known current page. Uses NavList's navigate_to_page for
+        navigation.
+
+        Args:
+            target_page (int): page to navigate to
+
+        Raises:
+            ValueError: invalid target_page specified
+        """
+        if target_page > self.ship_page_count:
+            raise ValueError(
+                "Invalid shiplist target page ({}) for number of known pages "
+                "({}).".format(target_page, self.ship_page_count))
+
+        self.current_shiplist_page = NavList.navigate_to_page(
+            self.regions, self.ship_page_count, self.current_shiplist_page,
+            target_page, 'repair')
+
     def _timer_to_datetime(self, timer):
         """Method to convert the passed in timer dict to a datetime object
 
@@ -226,12 +291,11 @@ class RepairModule(object):
         """Method to go through the existing repair timers and remove them if
         they are past, implying that the repairs are done and the dock is free
         """
-        now = datetime.now()
         self.repair_timers = [
-            timer for timer in self.repair_timers if timer > now]
+            timer for timer in self.repair_timers if timer > datetime.now()]
 
-    def _update_combat_next_sortie_time(self, timer):
-        """Method to update the combat module's next sortie time based on the
+    def _add_to_repair_timers(self, timer):
+        """Method to add to the internally tracked list of timers with the
         passed in timer.
 
         Args:
@@ -240,9 +304,18 @@ class RepairModule(object):
         """
         repair_end_time = self._timer_to_datetime(timer)
         self.repair_timers.append(repair_end_time)
+        Util.log_msg("Repair will complete at {}".format(
+            repair_end_time.strftime('%Y-%m-%d %H:%M:%S')))
 
-        if repair_end_time > self.combat.next_combat_time:
-            timer['minutes'] += 1
-            self.combat.set_next_combat_time(timer)
+    def _update_combat_next_sortie_time(self):
+        """Method to update the Combat module's next sortie timer based on the
+        shortest timer in the internal repair timers list.
+        """
+        self.repair_timers.sort()
+        shortest_timer = self.repair_timers[0]
+
+        if shortest_timer > self.combat.next_combat_time:
+            self.combat.next_combat_time = shortest_timer + timedelta(
+                minutes=1)
             Util.log_msg("Delaying next combat sortie to {}".format(
                 self.combat.next_combat_time.strftime('%Y-%m-%d %H:%M:%S')))