+++ CHANGELOG.md
@@ -1,3 +1,29 @@
+#### 0.5.1
+
+* Fix PortCheck on non-Event maps
+* Better checking of fatigue on ringed ships
+* Stability fix to live combat engine
+* Bugfix to sparkle mode in ShipSwitcher
+
+#### 0.5.0
+
+* Implements ScheduledStop functionality
+* Support fleet-specific ship repairs past the first page of the repair screen's ship list
+* Optimizations and bugfixes to node detection in live combat engine
+* Bugfix to ShipSwitcher being stuck on the last page
+* Bugfix to Expedition module crash when attempting to sortie a fleet with ships under repair
+
+#### 0.4.6-bugfix
+
+* Bugfix to repair module crashing if no timers are stored (thanks to @perryhuynh)
+
+#### 0.4.6
+
+* Improvements to setting of next combat timer based on repair timers in various situations
+* Implement daily quest reset
+* Force resupply after ship switching
+* Longer transient waits in Ship Switcher module
+
 #### 0.4.5
 
 * Archive 2018 Winter Event assets